move remove_config_response to model directory
