+++ lib/dart_remote_config.dart
@@ -1,3 +1,3 @@
 export 'model/remote_config.dart';
+export 'model/remote_config_response.dart';
 export 'remote_config_fetcher.dart';
-export 'remote_config_response.dart';
