+++ lib/model/remote_config_response.dart
