+++ test/remote_config_test.dart
@@ -1,7 +1,7 @@
 import 'package:dart_remote_config/model/promo_code.dart';
 import 'package:dart_remote_config/model/remote_config.dart';
+import 'package:dart_remote_config/model/remote_config_response.dart';
 import 'package:dart_remote_config/remote_config_fetcher.dart';
-import 'package:dart_remote_config/remote_config_response.dart';
 import 'package:test/test.dart';
 
 void main() {
